use customer id provided to `Braintree::PaymentMethod.create`
Ignored all parameters provided except the nonce. Only passed specs
because the customer id was included in the credit card hash passed
to the tokenize card method.

The customer id has been removed from the test hash and now all
parameters passed to .create are used in the creation of the credit
card.